+++ src/main/java/com/clubber/ClubberServer/domain/auth/component/UserRegister.java
@@ -2,6 +2,7 @@ package com.clubber.ClubberServer.domain.auth.component;
 
 import com.clubber.ClubberServer.domain.auth.dto.KakaoOauthResponse;
 import com.clubber.ClubberServer.domain.auth.service.AuthService;
+import com.clubber.ClubberServer.domain.auth.service.JwtTokenService;
 import com.clubber.ClubberServer.domain.user.domain.User;
 import com.clubber.ClubberServer.global.infrastructure.outer.api.oauth.client.kakao.KakaoInfoClient;
 import com.clubber.ClubberServer.global.infrastructure.outer.api.oauth.client.kakao.KakaoOauthClient;
@@ -16,19 +17,23 @@ import static com.clubber.ClubberServer.global.common.consts.ClubberStatic.BEARE
 @Component
 @RequiredArgsConstructor
 public class UserRegister {
-    private final AuthService authService;
-    private final KakaoOauthClient kakaoOauthClient;
-    private final KakaoInfoClient kakaoInfoClient;
-    private final KakaoProperties kakaoProperties;
 
-    public KakaoOauthResponse register(String code, String origin) {
-        KakaoTokenResponse kakaoTokenResponse = kakaoOauthClient.kakaoAuth(
-                kakaoProperties.getClientId(),
-                origin + kakaoProperties.getRedirectUrl(),
-                code);
+	private final AuthService authService;
+	private final JwtTokenService jwtTokenService;
+	private final KakaoOauthClient kakaoOauthClient;
+	private final KakaoInfoClient kakaoInfoClient;
+	private final KakaoProperties kakaoProperties;
 
-        KakaoUserInfoResponse kakaoUserInfoResponse = kakaoInfoClient.getUserInfo(BEARER + kakaoTokenResponse.getAccessToken());
-        User user = authService.loginOrSignUp(kakaoUserInfoResponse);
-        return authService.generateUserToken(user);
-    }
+	public KakaoOauthResponse register(String code, String origin) {
+		KakaoTokenResponse kakaoTokenResponse = kakaoOauthClient.kakaoAuth(
+			kakaoProperties.getClientId(),
+			origin + kakaoProperties.getRedirectUrl(),
+			code);
+
+		KakaoUserInfoResponse kakaoUserInfoResponse = kakaoInfoClient.getUserInfo(
+			BEARER + kakaoTokenResponse.getAccessToken());
+
+		User user = authService.loginOrSignUp(kakaoUserInfoResponse);
+		return jwtTokenService.generateUserToken(user);
+	}
 }
